Refactor to remove method fields in ParsedQObject
